+++ homeworks/vladyslav.kovalov_kovalov/7-oop/src/FullScreen.types.ts
@@ -1,0 +1,3 @@
+export interface IFullScreen {
+    toggle: (source: string) => void;
+}
