+++ homeworks/vladyslav.kovalov_kovalov/7-oop/src/ImageStorage.types.ts
@@ -1,0 +1,4 @@
+export interface IStorage {
+    images: string[];
+    save(image: string): number;
+}
