+++ homeworks/vladyslav.kovalov_kovalov/7-oop/src/Uploader.types.ts
@@ -1,0 +1,3 @@
+export interface IUploader {
+    listenEvents: (callback: CallableFunction) => void;
+}
